+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+* Issue **#17** : Added `include` and `exclude` functions.
+
 ## [v1.2.0-beta.1] - 2018-04-25
 
-* Issue **#15** : Added `if` and `not()` functions.
+* Issue **#15** : Added `if` and `not` functions.
 
 * Issue **#13** : Added value functions `true()`, `false()`, `null()` and `err()`.